requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
jupyter
notebook
astropy
polars
pyarrow
pyyaml
healpy
reproject
aplpy
pyvo
astroquery
huggingface_hub
joblib
tqdm
requests
optuna
python-dotenv
ydata_profiling